add validations to usuarioSistema

+++ backendfmo/src/main/java/com/backendfmo/controllers/UsuarioSistemaController.java
@@ -10,14 +10,15 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.backendfmo.dtos.request.usuariosistema.UsuarioSistemaDTO;
-import com.backendfmo.models.usuariosistema.UsuarioSistema;
+import com.backendfmo.repository.UsuarioRepository;
+import com.backendfmo.repository.UsuarioSistemaRepository;
 import com.backendfmo.services.usuarioSistema.IUsuarioSistemaService;
 
 import jakarta.validation.Valid;
-
 
 @RestController
 @CrossOrigin(origins = "*")
@@ -26,26 +27,36 @@ public class UsuarioSistemaController {
     @Autowired
     private IUsuarioSistemaService service;
 
+    @Autowired
+    private UsuarioSistemaRepository usuarioServicRepository;
 
     @GetMapping("/status")
-    public ResponseEntity<String> getStatus() {
-        return ResponseEntity.ok("API is running");
-    }
-
-    @GetMapping("/usuarioSistema")
-    public ResponseEntity<?> getAllUsuarioSistema(){
+    public ResponseEntity<String> getStatusAPI() {
         try {
-            return ResponseEntity.ok(service.getAllUsuarioSistema());
-            
+            return ResponseEntity.status(200).body("API is running");
         } catch (Exception e) {
             return ResponseEntity.status(500).body(e.getMessage());
         }
-        
+
     }
-   
+
+    @GetMapping("/usuarioSistema")
+    public ResponseEntity<?> getAllUsuarioSistema() {
+        try {
+            return ResponseEntity.status(202).body(service.getAllUsuarioSistema());
+        } catch (Exception e) {
+            return ResponseEntity.status(500).body(e.getMessage());
+        }
+
+    }
+
     @PostMapping("/crearUsuarioSistema")
     public ResponseEntity<?> createUsuarioSistema(@Valid @RequestBody UsuarioSistemaDTO usuarioSistema) {
-       try {
+        try {
+            boolean existeUsuario = usuarioServicRepository.existsByUsername(usuarioSistema.getUsername());
+            if (existeUsuario) {
+                return ResponseEntity.status(400).body("El usuario " + usuarioSistema.getUsername() + " ya existe");
+            }
             service.saveUsuarioSistema(usuarioSistema);
             URI location = service.createUri("/{id}", usuarioSistema);
             return (ResponseEntity<?>) ResponseEntity.created(location).build();
@@ -55,18 +66,26 @@ public class UsuarioSistemaController {
     }
 
     @GetMapping("/usuarioSistema/{id}")
-    public ResponseEntity<?> findUsuarioSistemaById(@Valid @PathVariable Integer id){
+    public ResponseEntity<?> findUsuarioSistemaById(@Valid @PathVariable Integer id) {
         try {
-            return ResponseEntity.ok(service.findUsuarioSistemaById(id));
+            return ResponseEntity.status(202).body(service.findUsuarioSistemaById(id));
         } catch (Exception e) {
             return ResponseEntity.status(500).body(e.getMessage());
         }
     }
 
     @DeleteMapping("/usuarioSistema/borrar/{username}")
-    public ResponseEntity<?> deleteUsuarioSistemaByUsername(@PathVariable String username){
-        service.deleteUsuarioSistema(username);
-        return ResponseEntity.ok().build();
+    public ResponseEntity<?> deleteUsuarioSistemaByUsername(@PathVariable String username) {
+        try {
+            boolean eliminado = usuarioServicRepository.existsByUsername(username);
+            if (!eliminado) {
+                return ResponseEntity.status(404).body("El usuario " + username + " no existe");
+            }
+            service.deleteUsuarioSistema(username);
+            return ResponseEntity.status(202).body("Usuario " + username + " eliminado correctamente");
+        } catch (Exception e) {
+            return ResponseEntity.status(500).body(e.getMessage());
+        }
     }
 
 }
